+++ src/promptflow/promptflow/_cli/_pf/_flow.py
@@ -29,7 +29,7 @@ from promptflow._cli._pf._init_entry_generators import (
     ChatFlowDAGGenerator,
     FlowDAGGenerator,
     OpenAIConnectionGenerator,
-    StreamlitFileGenerator,
+    StreamlitFileReplicator,
     ToolMetaGenerator,
     ToolPyGenerator,
     copy_extra_files,
@@ -389,10 +389,9 @@ def test_flow(args):
                 os.path.join(temp_dir, "logo.png"),
             ]
             for script in script_path:
-                StreamlitFileGenerator(
-                    flow_name=flow.name,
+                StreamlitFileReplicator(
+                    flow_name=flow.display_name if flow.display_name else flow.name,
                     flow_dag_path=flow.flow_dag_path,
-                    connection_provider=pf_client._ensure_connection_provider(),
                 ).generate_to_file(script)
             main_script_path = os.path.join(temp_dir, "main.py")
             pf_client.flows._chat_with_ui(script=main_script_path)