+++ src/promptflow/tests/sdk_cli_azure_test/unittests/test_cli.py
@@ -7,13 +7,12 @@ import pandas as pd
 import pytest
 from pytest_mock import MockFixture
 
-from promptflow._cli._pf_azure.entry import main
 from promptflow._sdk._constants import VIS_PORTAL_URL_TMPL
-from promptflow.azure.operations._flow_operations import FlowOperations
-from promptflow.azure.operations._run_operations import RunOperations
 
 
 def run_pf_command(*args, cwd=None):
+    from promptflow._cli._pf_azure.entry import main
+
     origin_argv, origin_cwd = sys.argv, os.path.abspath(os.curdir)
     try:
         sys.argv = ["pfazure"] + list(args)
@@ -46,6 +45,8 @@ class TestAzureCli:
         assert "0.0.1\n" in out
 
     def test_run_show(self, mocker: MockFixture, operation_scope_args):
+        from promptflow.azure.operations._run_operations import RunOperations
+
         mocked = mocker.patch.object(RunOperations, "get")
         # show_run will print the run object, so we need to mock the return value
         mocked.return_value._to_dict.return_value = {"name": "test_run"}
@@ -59,6 +60,8 @@ class TestAzureCli:
         mocked.assert_called_once()
 
     def test_run_show_details(self, mocker: MockFixture, operation_scope_args):
+        from promptflow.azure.operations._run_operations import RunOperations
+
         mocked = mocker.patch.object(RunOperations, "get_details")
         # show_run_details will print details, so we need to mock the return value
         mocked.return_value = pd.DataFrame([{"input": "input_value", "output": "output_value"}])
@@ -75,6 +78,8 @@ class TestAzureCli:
         mocked.assert_called_once()
 
     def test_run_show_metrics(self, mocker: MockFixture, operation_scope_args):
+        from promptflow.azure.operations._run_operations import RunOperations
+
         mocked = mocker.patch.object(RunOperations, "get_metrics")
         # show_metrics will print the metrics, so we need to mock the return value
         mocked.return_value = {"accuracy": 0.9}
@@ -95,6 +100,8 @@ class TestAzureCli:
         resource_group_name: str,
         workspace_name: str,
     ):
+        from promptflow.azure.operations._run_operations import RunOperations
+
         mocked_run = MagicMock()
         mocked_run._to_dict.return_value = {"name": "test_run"}
         mocked = mocker.patch.object(RunOperations, "list")
@@ -168,6 +175,8 @@ class TestAzureCli:
         mocker: MockFixture,
         operation_scope_args,
     ):
+        from promptflow.azure.operations._run_operations import RunOperations
+
         mocked = mocker.patch.object(RunOperations, "archive")
         mocked.return_value._to_dict.return_value = {"name": "test_run"}
         run_pf_command(
@@ -184,6 +193,8 @@ class TestAzureCli:
         mocker: MockFixture,
         operation_scope_args,
     ):
+        from promptflow.azure.operations._run_operations import RunOperations
+
         mocked = mocker.patch.object(RunOperations, "restore")
         mocked.return_value._to_dict.return_value = {"name": "test_run"}
         run_pf_command(
@@ -200,6 +211,8 @@ class TestAzureCli:
         mocker: MockFixture,
         operation_scope_args,
     ):
+        from promptflow.azure.operations._run_operations import RunOperations
+
         mocked = mocker.patch.object(RunOperations, "update")
         mocked.return_value._to_dict.return_value = {"name": "test_run"}
         run_pf_command(
@@ -220,6 +233,8 @@ class TestAzureCli:
         mocker: MockFixture,
         operation_scope_args,
     ):
+        from promptflow.azure.operations._flow_operations import FlowOperations
+
         mocked = mocker.patch.object(FlowOperations, "create_or_update")
         mocked.return_value._to_dict.return_value = {"name": "test_run"}
         run_pf_command(
@@ -241,6 +256,8 @@ class TestAzureCli:
         mocker: MockFixture,
         operation_scope_args,
     ):
+        from promptflow.azure.operations._flow_operations import FlowOperations
+
         mocked_flow = MagicMock()
         mocked_flow._to_dict.return_value = {"name": "test_flow"}
         mocked = mocker.patch.object(FlowOperations, "list")